max-bot/requirements.txt
aerich==0.8.2
aiofiles==25.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiomax==2.12.3
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.1.0
certifi==2024.12.14
cffi==1.17.1
cryptography==44.0.2
dictdiffer==0.9.0
dynaconf==3.2.10
frozenlist==1.5.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypika-tortoise==0.5.0
pytz==2025.1
sniffio==1.3.1
socksio==1.0.0
tenacity==9.0.0
tomli_w==1.2.0
tortoise-orm==0.24.2
typing_extensions==4.12.2
yarl==1.18.3
